Moving threshold classes from free-functions with static network_params initialization on to nano::work_thresholds class where they belong.
